Initialise PDF export page
