+++ packages/web-frontend/package.json
@@ -37,6 +37,7 @@
     "@material-ui/styles": "^4.9.10",
     "@tupaia/ui-components": "1.0",
     "@tupaia/utils": "1.0.0",
+    "connected-react-router": "^6.9.3",
     "core-js": "^3.16.1",
     "dom-to-image": "2.6.0",
     "dotenv": "^8.2.0",
@@ -59,7 +60,8 @@
     "react": "^16.13.1",
     "react-animations": "^1.0.0",
     "react-dom": "^16.13.1",
-    "react-redux": "^5.0.5",
+    "react-redux": "^7.0.0",
+    "react-router-dom": "^6.3.0",
     "react-tap-event-plugin": "^3.0.3",
     "recharts": "^1.8.5",
     "redux": "^3.6.0",
