+++ packages/web-frontend/src/screens/desktop/RootScreen/PDFExportPage.js
@@ -1,0 +1,18 @@
+import React from 'react';
+import PropTypes from 'prop-types';
+import { connect } from 'react-redux';
+import styled from 'styled-components';
+
+const Main = styled.div`
+  height: 100vh;
+  background: white;
+`;
+const PDFExportPage = () => <Main>123</Main>;
+
+PDFExportPage.propTypes = {};
+
+PDFExportPage.defaultProps = {};
+
+const mapStateToProps = state => {};
+
+export default connect(mapStateToProps)(PDFExportPage);
